pyproject.toml
[build-system]
requires = ["flit_core>=3.4"]
build-backend = "flit_core.buildapi"

[project]
name = "hquant"
readme = "README.md"
requires-python = ">=3.8"
license = { file="LICENSE" }
authors = [
	{ name="Marcos Del Sol Vives", email="marcos@orca.pet" },
]
keywords = ["image", "dither", "dithering", "quantization", "palette"]
classifiers = [
	"License :: Freely Distributable",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: 3.9",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Topic :: Multimedia",
	"Topic :: Multimedia :: Graphics",
	"Topic :: Multimedia :: Graphics :: Graphics Conversion",
	"Typing :: Typed",
]
dependencies = [
    "Pillow >=9.0.0",
]
dynamic = ["version", "description"]

[project.urls]
Homepage = "https://github.com/socram8888/hquant"
Issues = "https://github.com/socram8888/hquant/issues"
